[windows] fix for x64

+++ src/Pyro.cpp
@@ -171,7 +171,7 @@ extern "C" void Render()
 {
 #ifdef WIN32
   g_pContext->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLESTRIP);
-  size_t strides = sizeof(MYCUSTOMVERTEX), offsets = 0;
+  UINT strides = sizeof(MYCUSTOMVERTEX), offsets = 0;
   g_pContext->IASetVertexBuffers(0, 1, &g_pVBuffer, &strides, &offsets);
   g_pContext->PSSetShader(g_pPShader, NULL, 0);
 #endif
